+++ flutter_value_bloc/example/lib/list/ListNameCubit.dart
@@ -5,7 +5,6 @@ class ListPersonCubit extends ListValueCubit<Person, Object> {
   @override
   void onFetching(FetchScheme scheme) async {
     await Future.delayed(Duration(seconds: 2));
-    print(scheme);
     emitFetched(scheme, personList.skip(scheme.offset).take(scheme.limit));
   }
 }