+++ value_bloc/lib/src/ValueCubitObserver.dart
@@ -1,9 +1,12 @@
-import 'package:value_bloc/value_bloc.dart';
+import '../value_bloc.dart';
 
 class ValueCubitObserver {
+  static bool isEnabledPrintMethodIgnored = false;
+
   static ValueCubitObserver instance = ValueCubitObserver();
 
   void methodIgnored(ValueState<dynamic> state, String nameMethod) {
+    if (!ValueCubitObserver.isEnabledPrintMethodIgnored) return;
     print('method:$nameMethod,state:$state');
     print(StackTrace.current);
   }