+++ value_bloc/CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - TODO: Add release date.
+## 0.1.0
+- Converted the blocs to cubits
+- Converted the Status Enums in cubit state
+- Implemented a custom Fetchers
+- More improvements
 
-* TODO: Describe initial release.
+## 0.0.1
+InitialRelease